requirements.txt
jinja2>=3.1.0
weasyprint>=60.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
feedparser>=6.0.0
selenium>=4.0.0
docling>=0.1.0
python-dotenv>=1.0.1
python-dateutil>=2.8.0